feat: new coupon factory with refactorized coupon methods from bond
