+++ packages/ats/sdk/src/app/usecase/query/bond/coupons/getCouponFor/GetCouponForQueryHandler.unit.test.ts
@@ -6,11 +6,12 @@ import { ErrorCode } from "@core/error/BaseError";
 import { RPCQueryAdapter } from "@port/out/rpc/RPCQueryAdapter";
 import ContractService from "@service/contract/ContractService";
 import EvmAddress from "@domain/context/contract/EvmAddress";
-import { GetCouponForQueryFixture } from "@test/fixtures/bond/BondFixture";
+import { CouponFixture, GetCouponForQueryFixture } from "@test/fixtures/bond/BondFixture";
 import { GetCouponForQueryError } from "./error/GetCouponForQueryError";
 import { GetCouponForQueryHandler } from "./GetCouponForQueryHandler";
 import AccountService from "@service/account/AccountService";
 import BigDecimal from "@domain/context/shared/BigDecimal";
+import { CouponAmountFor } from "@domain/context/bond/CouponAmountFor";
 import { GetCouponForQuery, GetCouponForQueryResponse } from "./GetCouponForQuery";
 
 describe("GetCouponForQueryHandler", () => {
@@ -26,7 +27,10 @@ describe("GetCouponForQueryHandler", () => {
 
   const errorMsg = ErrorMsgFixture.create().msg;
   const amount = new BigDecimal("1000000");
+  const nominalValue = new BigDecimal("500");
   const decimals = 6;
+  const coupon = CouponFixture.create();
+  const couponAmount = new CouponAmountFor("10", "4", true);
 
   beforeEach(() => {
     handler = new GetCouponForQueryHandler(queryAdapterServiceMock, accountServiceMock, contractServiceMock);
@@ -57,7 +61,11 @@ describe("GetCouponForQueryHandler", () => {
       accountServiceMock.getAccountEvmAddress.mockResolvedValueOnce(targetEvmAddress);
       queryAdapterServiceMock.getCouponFor.mockResolvedValue({
         tokenBalance: amount,
+        nominalValue: nominalValue,
         decimals: decimals,
+        recordDateReached: false,
+        coupon: coupon,
+        couponAmount: couponAmount,
         isDisabled: false,
       });
 
@@ -65,7 +73,11 @@ describe("GetCouponForQueryHandler", () => {
 
       expect(result).toBeInstanceOf(GetCouponForQueryResponse);
       expect(result.tokenBalance).toStrictEqual(amount);
+      expect(result.nominalValue).toStrictEqual(nominalValue);
       expect(result.decimals).toStrictEqual(decimals);
+      expect(result.recordDateReached).toBe(false);
+      expect(result.coupon).toStrictEqual(coupon);
+      expect(result.couponAmount).toStrictEqual(couponAmount);
       expect(result.isDisabled).toBe(false);
       expect(contractServiceMock.getContractEvmAddress).toHaveBeenCalledTimes(1);
       expect(accountServiceMock.getAccountEvmAddress).toHaveBeenCalledTimes(1);
